Refactor code editor components and enhance expression handling
- Replace CodeMirrorEditor with CodeEditor across multiple files for consistency.
- Introduce YamlStyledEditor and JsonStyledEditor for improved YAML and JSON editing capabilities.
- Implement ForEachField component to manage multiple expressions in control flow.
- Enhance expression input with better completion and tooltip support.
- Remove deprecated JsonEditor component and related code.
- Add utility functions for template expression handling and error reporting.
